Removed Windows only HiRes timer
Now using the timer functionality provded by GLFW for multiplatform
compatability

+++ OpenGLT1/OpenGLT1/OpenGLSys.cpp
@@ -68,7 +68,7 @@ void OpenGLSys::Disable2D()
 	glMatrixMode(GL_MODELVIEW);
 	glPopMatrix();
 }
-void OpenGLSys::Update(float dt)
+void OpenGLSys::Update(double dt)
 {
 	m_rotation += dt * 50;
 }
